Fix free shipping bar update and threshold
Refactored `nraizes_free_shipping_bar` to use AJAX fragments for dynamic updates.
- Set free shipping threshold to R$ 500.
- Implemented `nraizes_get_free_shipping_html` helper.
- Added AJAX fragment refresh for cart and checkout updates.
- Used `(float) WC()->cart->subtotal` for calculation safety.
- Added debug info to HTML comment.

+++ wp-content/themes/organium-child/inc/checkout.php
@@ -47,12 +47,13 @@ function nraizes_get_free_shipping_html() {
     }
 
     $min_amount = 500;
-    $current = WC()->cart->subtotal;
+    $current = (float) WC()->cart->subtotal;
     $remaining = $min_amount - $current;
     
     ob_start();
     ?>
     <div id="nraizes-free-shipping-bar-wrapper">
+        <!-- Debug: Current: <?php echo $current; ?> | Min: <?php echo $min_amount; ?> | Remaining: <?php echo $remaining; ?> -->
         <?php if ($remaining > 0) :
             $percent = ($current / $min_amount) * 100;
             ?>
